Clean-up: unused imports, commented code, type annotations, etc
